requirements.txt
aiohttp
nest_asyncio
python-dotenv
colorama
graphviz
IPython
beautifulsoup4
tiktoken
numpy
fpdf
